[sec] ensure that nobody can go outside cache, tmp & log folders through adapter
