+++ tests/Unit/Adapter/JoomlaFolderTest.php
@@ -9,12 +9,13 @@
 
 namespace Phproberto\Joomla\Flysystem\Tests\Unit\Adapter;
 
+use LogicException;
 use Joomla\CMS\Factory;
 use League\Flysystem\AdapterInterface;
 use Joomla\CMS\Application\CMSApplication;
 use Phproberto\Joomla\Flysystem\Adapter\JoomlaFolder;
 use Phproberto\Joomla\Flysystem\Tests\Unit\TestWithEvents;
-use LogicException;
+use Phproberto\Joomla\Flysystem\Tests\Unit\Adapter\Traits\IsBlockedInsideRoot;
 
 /**
  * JoomlaFolder adapter tests.
@@ -23,6 +24,8 @@ use LogicException;
  */
 class JoomlaFolderTest extends TestWithEvents
 {
+	use IsBlockedInsideRoot;
+
 	/**
 	 * Tested adapter.
 	 *
@@ -108,20 +111,6 @@ class JoomlaFolderTest extends TestWithEvents
 	}
 
 	/**
-	 * @test
-	 *
-	 * @dataProvider insecurePaths
-	 *
-	 * @return void
-	 *
-	 * @expectedException  LogicException
-	 */
-	public function throwsExceptonOnInsecurePaths($path)
-	{
-		$this->adapter->has($path);
-	}
-
-	/**
 	 * Triggered before adapter has been loaded.
 	 *
 	 * @param   AdapterInterface  $adapter  Adapter being instatiated
@@ -181,20 +170,4 @@ class JoomlaFolderTest extends TestWithEvents
 
 		$adapter->setParam('onFlysystemAfterLoadJoomlaFolderAdapter', true);
 	}
-
-	/**
-	 * List of insecure paths to prevent unauthorised access.
-	 *
-	 * @return  array
-	 */
-	public function insecurePaths()
-	{
-		return [
-			['../index.php'],
-			['./../index.php'],
-			['./jui/../../index.php'],
-			['..']
-		];
-	}
-
 }
